add tests for VirtualMachineOperandStack support
create a methodbuilder that uses operand stack operations with occasional
commit calls directly, calling native functions to verify that the state
of the stack is only reflected after a Commit() call, but that computations
can be correctly made with the simulated operand stack.

This first version of the test uses an operand stack that consists of
32-bit integers and the stack pointer is maintained as an index into
an array of 32-bit integers.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>