Add operand stack test using VirtualMachineRegisterInStruct and fix problems
OperandStackTests should really test using both VirtualMachineRegister as well
as VirtualMachineRegisterInStruct. Refactored so that most of the actual test
and verification code is shared between these two tests. The InStruct version
of the test compiles a method that takes a Thread structure which has a field
containing the stack pointer. Since nothing has used VirtualMachineRegisterInStruct
before these tests (and since it's just a header file) there were quite a few
problems that needed to be fixed.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>