Fix virtual machine state management for merges and add test to OperandStackTests
The current operand stack test only ensures that the operand stack is properly
modelled on a single control flow path. I added several BytecodeBuilder objects
at the end of the current tests to create an if/then/else diamond and verified
that values Pushed from either side of the diamond reach the merge point. The
version of the tests that uses a straight pointer exercises the "then" path of
the generated code. The version of the tests that uses a stack pointer embedded
in a "thread" structure exercises the "else" path of the generated code.
Technically, these paths are tested in different method bodies and it wouuld
probably be better to test both paths in each generated body, but since the
code to generate the if/then/else diamond is shared, it should providing the
same level of testing.

Because the handling of branches became more complicated, I have now created
version of Goto and IfCmp* on OMR::BytecodeBuilder itself, which removes the
need to call AddSuccessorBuilder() directly. There is a bit of a ballet
required when creating branches between BytecodeBuilders now because of the
potential need to merge the vm state information. If merging is required,
an intermediate BytecodeBuilder object is introduced along the flow edge
being created, and any operations needed to synchronize the vm state at the
target are placed into this intermediate builder object. The complication
that process introduces is that the client asked for control to flow to
a particular place, but the actual operation that performs that control
flow may need to be directed towards the intermediate builder. Rather than
impose this subtlety on the clients, I implemented it all inside the
control flow services the client is already calling (see, for example,
OMR::BytecodeBuilder::Goto(TR::BytecodeBuilder **dest) ).

I also made some tracing improvements for IfCmp* opcodes to make it
easier to see the various control flow connections between builder
objects.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>