Add bytecode index worklist service to MethodBuilder
The automatic propagation of virtual machine state variables performed
by BytecodeBuilders requires that the language compiler visit bytecodes
in a particular order. In particular, no bytecode should be visited
before at least one of its predecessors has been visited. This
requirement stems from the fact that it is the creation of that
flow edge from the predecessor to the bytecode that causes the
vm state to propagate. If you visit a bytecode before any of its
predecessors, it will almost certainly have a NULL vmState.

To facilitate the correct traversal of the bytecodes, I'm going
to add a worklist facility to MethodBuilder, which will keep
track of which bytecode indices have already been visited and
which bytecode indices are ready to be visited. A pair of bit
vectors serves the purpose nicely. Then, by augmenting the
services in MethodBuilder and BytecodeBuilder that create control
flow edges between BytecodeBuilder objects, the insertion of
bytecode indices to the worklist can be done automatically.

The language compiler must create an initial VM state and
explicitly store it to the MethodBuilder object. The first
control flow edge from the MethodBuilder to the first
bytecode (typically at bytecode index 0) must be created by
calling AppendBuilder() on the MethodBuilder. From there, the
language compiler need only iterate calling
GetNextBytecodeFromWorklist() to get the next bytecode index
to be processed until it returns -1. When -1 is returned,
there are no remaining bytecode builders to visit. If there
are any other special entry points, then the bytecode builders
for those special entry points will need to be explicitly added
to the worklist by passing them to addBytecodeBuilderToWorklist().

This commit also adds a test of the worklist, which verifies that
the reachable bytecode builders and only those builders are
traversed by the worklist. The method will fail to compile if
the worklist does not visit exactly the right builders in
exactly the expected (from lowest to highest bc index ) order.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>