JitBuilder Goto should not always mark builder as not coming back
There is a Goto appended to a builder in one of the looping services
that is actually in a conditional path, so the builder object should
not be marked as not coming back (i.e. it needs no fall-through edge).
Fixed by specifically marking the builder object in this case as coming
back (so Goto by default still marks the builder object as not coming
back).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>