Add InlineCall service to JitBuilder
Rudimentary inliner for calls to a target represented by a
MethodBuilder object (looking up the right target is up to
the client, but since all languages do this part
differently...). Basic approach is to first insert the
code to store the arguments to the incoming parameter
names, then use AppendBuilder() to place the inlined
code body. A return block and return value are created
and passed to the inlined method builder object.

Return is modified so that, in the case of an inlined
method builder object, it will route to the provided
return block. Any return value is written into the
provided return value.

Symbol lookup in MethodBuilder is made virtual, so
that subclasses can decide how and whether to delegate
symbol lookup from the inlined body to the caller.

This commit does not deal with several complexities,
including exception handling / edge hookup, the
fact that symbols in the inlined code body may
conflict with symbol names in the caller (particularly
dangerous for parameter names), or the handling of
vm state (though the inlined body has an independent
vm state from the caller).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>