Improve static block frequency prediction
In the absence of profile data, the compiler needs to predict block
frequencies. From its Java heritage, the optimizer has been well
trained to exploit block frequencies, so not having good block
frequencies can drive all kinds of nonsensible behaviour. This
commit attempts to reboot the static frequency prediction with a
simpler but more sensible default block frequency assignment
just based on loop depth. This approach at least points the
optimizer in a sensible direction. It does not attempt to
predict branch directions, which means optimizations like
block ordering tend to maintain the initial ordering of
blocks when loop depth is the same.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>