Fix block freq distribution bug in OrderBlocks
One peephole optimization in OrderBlocks tries to
change a conditional branch to a block containing just
a goto so that the conditional branch goes directly to
the destination of the goto. In the process, it tries
to update block frequencies but mistakenly double
counts one edge's frequency at the target block for
the goto.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>