Use createNamedSymbol for TypeDictionary union fields
TypeDictionary currently creates symbols for union fields using
createSymbol which does not take a name parameter. So in JIT logs,
the symbols all come out as "unknown field". This commit changes
TypeDictionary to use createNamedSymbol instead to pass in the actual
name of the field so it shows up in the logs properly as
"union.field"

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>