Make ELF file under perfTool an object file
Adjust the parameters to make the generated object file when
running with TR_Options="perfTool" to be an object file rather
than an executable. May no longer be able to use the resulting
ELF file as an input to objdump for annotation with perf samples.
Add relocation section, but currently empty.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>