Eliminate early jitbuilder replay support
The file ReplayMethod.cpp is a remnant of an earlier prototype for replay
compilation for JitBuilder, along with a few targets in release/Makefile.
There are also macros defined and called by IlBuilder and MethodBuilder
that are about to be replaced by a better implementation of JBIL recording/
replaying.

These can all be eliminated.

Issue: #1818
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>