Define MethodBuilderAPI class
This commit creates a new (extensible) base class for the MethodBuilder
class that specfically defines the public (client) API for the MethodBuilder
part of the JitBuilder API. This class should never have any protected or
private API or implementation; the MethodBuilderAPI class is an interface.
The implementation of that interface is currently provided by MethodBuilder
itself, but alternative implementations (say, one that simple records the
API calls) can be created in future.

Once all classes have a public API defined, we can introduce a versioning
scheme that can capture how the backwards compatibility of future API changes
evolves over time.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>