Make MethodBuilder into a more extensible class
In preparation for creating a JitBuilder extension for MethodBuilder,
to facilitate the client API creation, the MethodBuilder class needs
to become more like a fully extensible with an independent header.

Because the test compiler provides an extension of MethodBuilder,
its files were also adjusted to define a separate header file.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>