Add client object pointers to JitBuilder base classes
As part of the enablement of the JitBuilder client API, the notion
that client objects which are 1-1 mapped to JitBuilder objects is
introduced. On the JitBuilder side, this change introduces a field
to all JitBuilder objects to hold a void * pointer to this client
object. Note that the client object may ultimately be a value that
only has meaning in some language runtime.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>