Add IlBuilder buildIL client callback support
Defining the client API for JitBuilder includes providing a mechanism for
all the call backs used by JitBuilder to be used by clients via the
client API. This mechanism is a C callback for all virtual functions
that can be overidden. For IlBuilder, that is buildIL() which returns
a bool. The callback can be set via setClientCallback_buildIL() and
will be called with the client object as parameter.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>