Fix up Switch API
Using "bool" as a vararg is actually undefined behaviour, so change
that type to an int32_t. Also, with the new client API, the vararg
implementation is no longer required because the client API will only
call the "array" version. Nonetheless, since it's fed by the client
API, the type for the "falls through" condition array needs to become
int32_t.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>